feat: Refatora configurações de relacionamentos do EF Core

+++ ApiBiblioteca.Application/DTOs/DtosExemplar/CreateExemplarDto.cs
@@ -7,5 +7,5 @@ public class CreateExemplarDto
     public string Nome { get; set; }
     public string CodigoDeBarras { get; set; }
     public decimal Preco { get; set; }
-    public long LivroId { get; set; }
+    public int LivroId { get; set; }
 }
