+++ ApiBiblioteca.Application/DTOs/DtosLivro/CreateLivroDto.cs
@@ -8,6 +8,6 @@ public class CreateLivroDto
     public DateOnly DataPublicacao { get; set; }
     public int NumeroDePaginas { get; set; }
     public string Isbn { get; set; }
-    public long CategoriaId { get; set; }
-    public long AutorId { get; set; }
+    public int CategoriaId { get; set; }
+    public int AutorId { get; set; }
 }
