+++ ApiBiblioteca.Application/DTOs/DtosLivro/LivroResponseDto.cs
@@ -2,12 +2,12 @@
 
 public class LivroResponseDto()
 {
-    public long Id { get; set; } 
+    public int Id { get; set; } 
     public string Titulo { get; set; } 
     public DateOnly DataPublicacao { get; set; } 
     public DateOnly DataCadastro { get; set; }
     public int NumeroDePaginas { get; set; } 
     public string Isbn { get; set; } 
-    public long CategoriaId { get; set; } 
-    public long AutorId { get; set; } 
+    public int CategoriaId { get; set; } 
+    public int AutorId { get; set; } 
 }
